+++ src/autoscaling/snapshotter.js
@@ -1,3 +1,4 @@
+import os from 'os';
 import _ from 'underscore';
 import { betterSetInterval, betterClearInterval } from 'apify-shared/utilities';
 import log from 'apify-shared/log';
@@ -11,11 +12,14 @@ const DEFAULT_OPTIONS = {
     maxBlockedMillis: 50, // 0.05
     memorySnapshotIntervalSecs: 1,
     clientSnapshotIntervalSecs: 1,
+    cpuSnapshotIntervalSecs: 1,
     maxUsedMemoryRatio: 0.7,
+    maxUsedCpuRatio: 0.95,
     snapshotHistorySecs: 30,
     maxClientErrors: 3,
 };
 const RESERVE_MEMORY_RATIO = 0.5;
+const CLIENT_RATE_LIMIT_ERROR_RETRY_COUNT = 2;
 
 /**
  * Creates snapshots of system resources at given intervals and marks the resource
@@ -23,21 +27,23 @@ const RESERVE_MEMORY_RATIO = 0.5;
  * It tracks the following resources: Memory, EventLoop, API and CPU.
  * The class is used by the {@link AutoscaledPool} class.
  *
- * There are differences in behavior when running locally and on the Apify platform,
- * but those differences are handled internally by the class and do not affect its interface.
+ * When running on the Apify platform, the CPU and memory statistics are provided by the platform,
+ * as collected from the running Docker container. When running locally, `Snapshotter`
+ * makes its own statistics by querying the OS.
+ *
+ * CPU becomes overloaded locally when its current use exceeds the `maxUsedCpuRatio` option or
+ * when Apify platform marks it as overloaded.
  *
  * Memory becomes overloaded if its current use exceeds the `maxUsedMemoryRatio` option.
  * It's computed using the total memory available to the container when running on
  * the Apify platform and a quarter of total system memory when running locally.
- * Max total memory may be overridden by using the `APIFY_MEMORY_MBYTES` environment variable.
+ * Max total memory when running locally may be overridden by using the `APIFY_MEMORY_MBYTES`
+ * environment variable.
  *
  * Event loop becomes overloaded if it slows down by more than the `maxBlockedMillis` option.
  *
  * Client becomes overloaded when rate limit errors (429 - Too Many Requests),
- * typically received from the request queue exceed the set limit within the set interval.
- *
- * CPU tracking is available only on the Apify platform and the CPU overloaded event is read
- * directly off the container and is not configurable.
+ * typically received from the request queue, exceed the set limit within the set interval.
  *
  * @param {Object} [options] All `Snapshotter` parameters are passed
  *   via an options object with the following keys:
@@ -49,8 +55,17 @@ const RESERVE_MEMORY_RATIO = 0.5;
  * @param {Number} [options.maxBlockedMillis=50]
  *   Maximum allowed delay of the event loop in milliseconds.
  *   Exceeding this limit overloads the event loop.
+ * @param {Number} [options.cpuSnapshotIntervalSecs=1]
+ *   Defines the interval of measuring CPU usage.
+ *   This is only used when running locally. On the Apify platform,
+ *   the statistics are provided externally at a fixed interval.
+ * @param {Number} [options.maxUsedCpuRatio=0.95]
+ *   Defines the maximum usage of CPU.
+ *   Exceeding this limit overloads the CPU.
  * @param {Number} [options.memorySnapshotIntervalSecs=1]
  *   Defines the interval of measuring memory consumption.
+ *   This is only used when running locally. On the Apify platform,
+ *   the statistics are provided externally at a fixed interval.
  *   The measurement itself is resource intensive (25 - 50ms async).
  *   Therefore, setting this interval below 1 second is not recommended.
  * @param {Number} [options.maxUsedMemoryRatio=0.7]
@@ -67,29 +82,35 @@ class Snapshotter {
     constructor(options = {}) {
         const {
             eventLoopSnapshotIntervalSecs,
+            cpuSnapshotIntervalSecs,
             memorySnapshotIntervalSecs,
             clientSnapshotIntervalSecs,
             snapshotHistorySecs,
             maxBlockedMillis,
             maxUsedMemoryRatio,
+            maxUsedCpuRatio,
             maxClientErrors,
         } = _.defaults({}, options, DEFAULT_OPTIONS);
 
         checkParamOrThrow(eventLoopSnapshotIntervalSecs, 'options.eventLoopSnapshotIntervalSecs', 'Number');
         checkParamOrThrow(memorySnapshotIntervalSecs, 'options.memorySnapshotIntervalSecs', 'Number');
+        checkParamOrThrow(cpuSnapshotIntervalSecs, 'options.cpuSnapshotIntervalSecs', 'Number');
         checkParamOrThrow(snapshotHistorySecs, 'options.snapshotHistorySecs', 'Number');
         checkParamOrThrow(clientSnapshotIntervalSecs, 'options.clientSnapshotIntervalSecs', 'Number');
         checkParamOrThrow(maxBlockedMillis, 'options.maxBlockedMillis', 'Number');
         checkParamOrThrow(maxUsedMemoryRatio, 'options.maxUsedMemoryRatio', 'Number');
+        checkParamOrThrow(maxUsedCpuRatio, 'options.maxUsedCpuRatio', 'Number');
         checkParamOrThrow(maxClientErrors, 'options.maxClientErrors', 'Number');
 
 
         this.eventLoopSnapshotIntervalMillis = eventLoopSnapshotIntervalSecs * 1000;
         this.memorySnapshotIntervalMillis = memorySnapshotIntervalSecs * 1000;
         this.clientSnapshotIntervalMillis = clientSnapshotIntervalSecs * 1000;
+        this.cpuSnapshotIntervalMillis = cpuSnapshotIntervalSecs * 1000;
         this.snapshotHistoryMillis = snapshotHistorySecs * 1000;
         this.maxBlockedMillis = maxBlockedMillis;
         this.maxUsedMemoryRatio = maxUsedMemoryRatio;
+        this.maxUsedCpuRatio = maxUsedCpuRatio;
         this.maxClientErrors = maxClientErrors;
         this.maxMemoryBytes = (parseInt(process.env[ENV_VARS.MEMORY_MBYTES], 10) * 1024 * 1024) || null;
 
@@ -101,6 +122,7 @@ class Snapshotter {
         this.eventLoopInterval = null;
         this.memoryInterval = null;
         this.clientInterval = null;
+        this.cpuInterval = null;
     }
 
     /**
@@ -108,23 +130,18 @@ class Snapshotter {
      * @return {Promise}
      */
     async start() {
-        // Ensure max memory is correctly computed.
-        if (!this.maxMemoryBytes) {
-            const { totalBytes } = await getMemoryInfo();
-            if (isAtHome()) {
-                this.maxMemoryBytes = totalBytes;
-            } else {
-                this.maxMemoryBytes = Math.ceil(totalBytes / 4);
-                // NOTE: Log as AutoscaledPool, so that users are not confused what "Snapshotter" is
-                log.info(`AutoscaledPool: Setting max memory of this run to ${Math.round(this.maxMemoryBytes / 1024 / 1024)} MB. Use the ${ENV_VARS.MEMORY_MBYTES} environment variable to override it.`); // eslint-disable-line max-len
-            }
-        }
+        await this._ensureCorrectMaxMemory();
 
         // Start snapshotting.
         this.eventLoopInterval = betterSetInterval(this._snapshotEventLoop.bind(this), this.eventLoopSnapshotIntervalMillis);
-        this.memoryInterval = betterSetInterval(this._snapshotMemory.bind(this), this.memorySnapshotIntervalMillis);
         this.clientInterval = betterSetInterval(this._snapshotClient.bind(this), this.clientSnapshotIntervalMillis);
-        events.on(ACTOR_EVENT_NAMES.CPU_INFO, this._snapshotCpu.bind(this));
+        if (isAtHome()) {
+            events.on(ACTOR_EVENT_NAMES.SYSTEM_INFO, this._snapshotCpuOnPlatform.bind(this));
+            events.on(ACTOR_EVENT_NAMES.SYSTEM_INFO, this._snapshotMemoryOnPlatform.bind(this));
+        } else {
+            this.cpuInterval = betterSetInterval(this._snapshotCpuOnLocal.bind(this), this.cpuSnapshotIntervalMillis);
+            this.memoryInterval = betterSetInterval(this._snapshotMemoryOnLocal.bind(this), this.memorySnapshotIntervalMillis);
+        }
     }
 
     /**
@@ -134,8 +151,10 @@ class Snapshotter {
     async stop() {
         betterClearInterval(this.eventLoopInterval);
         betterClearInterval(this.memoryInterval);
+        betterClearInterval(this.cpuInterval);
         betterClearInterval(this.clientInterval);
-        events.removeListener(ACTOR_EVENT_NAMES.CPU_INFO, this._snapshotCpu);
+        events.removeListener(ACTOR_EVENT_NAMES.SYSTEM_INFO, this._snapshotCpuOnPlatform);
+        events.removeListener(ACTOR_EVENT_NAMES.SYSTEM_INFO, this._snapshotMemoryOnPlatform);
         // Allow microtask queue to unwind before stop returns.
         await new Promise(resolve => setImmediate(resolve));
     }
@@ -204,12 +223,33 @@ class Snapshotter {
     }
 
     /**
-     * Creates a snapshot of current memory usage.
+     * Creates a snapshot of current memory usage
+     * using the Apify platform `systemInfo` event.
+     * @param {Object} systemInfo
+     * @ignore
+     */
+    _snapshotMemoryOnPlatform(systemInfo) {
+        const now = new Date();
+        this._pruneSnapshots(this.memorySnapshots, now);
+        const { memCurrentBytes } = systemInfo;
+        const snapshot = {
+            createdAt: now,
+            isOverloaded: memCurrentBytes / this.maxMemoryBytes > this.maxUsedMemoryRatio,
+            usedBytes: memCurrentBytes,
+        };
+
+        this.memorySnapshots.push(snapshot);
+        this._memoryOverloadWarning(systemInfo);
+    }
+
+    /**
+     * Creates a snapshot of current memory usage
+     * using the Apify platform `systemInfo` event.
      * @param {Function} intervalCallback
      * @return {Promise}
      * @ignore
      */
-    async _snapshotMemory(intervalCallback) {
+    async _snapshotMemoryOnLocal(intervalCallback) {
         try {
             const now = new Date();
             const memInfo = await getMemoryInfo();
@@ -217,15 +257,13 @@ class Snapshotter {
             this._pruneSnapshots(this.memorySnapshots, now);
 
             const usedBytes = mainProcessBytes + childProcessesBytes;
-            const snapshot = Object.assign(memInfo, {
+            const snapshot = {
                 createdAt: now,
                 isOverloaded: usedBytes / this.maxMemoryBytes > this.maxUsedMemoryRatio,
-            });
+                usedBytes,
+            };
 
             this.memorySnapshots.push(snapshot);
-            // TODO revisit this once we add memory readings from platform.
-            // It only scares people and provides no value, because it's always wrong about memory.
-            // this._memoryOverloadWarning(memInfo);
         } catch (err) {
             log.exception(err, 'Snapshotter: Memory snapshot failed.');
         } finally {
@@ -236,23 +274,18 @@ class Snapshotter {
     /**
      * Checks for critical memory overload and logs it to the console.
      * @ignore
-     * @param {Object} memoryInfo - Memory information
+     * @param {Object} systemInfo
      */
-    _memoryOverloadWarning(memoryInfo) {
-        try {
-            const maxDesiredMemoryBytes = this.maxUsedMemoryRatio * this.maxMemoryBytes;
-            const reserveMemory = this.maxMemoryBytes * (1 - this.maxUsedMemoryRatio) * RESERVE_MEMORY_RATIO;
-            const criticalOverloadBytes = maxDesiredMemoryBytes + reserveMemory;
-            const { mainProcessBytes, childProcessesBytes } = memoryInfo;
-            const usedBytes = mainProcessBytes + childProcessesBytes;
-            const isCriticalOverload = usedBytes > criticalOverloadBytes;
-            if (isCriticalOverload) {
-                const usedPercentage = usedBytes / this.maxMemoryBytes * 100;
-                const toMb = bytes => Math.round(bytes / (1024 ** 2));
-                log.warning(`Memory is critically overloaded. Used ${toMb(usedBytes)} MB of ${toMb(this.maxMemoryBytes)} MB (${usedPercentage}%)`);
-            }
-        } catch (e) {
-            log.exception(e, 'Snapshotter: Memory leak warning failed.');
+    _memoryOverloadWarning({ memCurrentBytes }) {
+        const maxDesiredMemoryBytes = this.maxUsedMemoryRatio * this.maxMemoryBytes;
+        const reserveMemory = this.maxMemoryBytes * (1 - this.maxUsedMemoryRatio) * RESERVE_MEMORY_RATIO;
+        const criticalOverloadBytes = maxDesiredMemoryBytes + reserveMemory;
+        const isCriticalOverload = memCurrentBytes > criticalOverloadBytes;
+        if (isCriticalOverload) {
+            const usedPercentage = memCurrentBytes / this.maxMemoryBytes * 100;
+            const toMb = bytes => Math.round(bytes / (1024 ** 2));
+            log.warning('Memory is critically overloaded. '
+                + `Used ${toMb(memCurrentBytes)} MB of ${toMb(this.maxMemoryBytes)} MB (${usedPercentage}%)`);
         }
     }
 
@@ -262,45 +295,88 @@ class Snapshotter {
      * @ignore
      */
     _snapshotEventLoop(intervalCallback) {
-        try {
-            const now = new Date();
-            this._pruneSnapshots(this.eventLoopSnapshots, now);
+        const now = new Date();
+        this._pruneSnapshots(this.eventLoopSnapshots, now);
 
-            // Handle empty snapshots array
-            const snapshot = {
-                createdAt: now,
-                isOverloaded: false,
-                exceededMillis: 0,
-            };
-            const previousSnapshot = this.eventLoopSnapshots[this.eventLoopSnapshots.length - 1];
-            if (!previousSnapshot) return this.eventLoopSnapshots.push(snapshot);
+        const snapshot = {
+            createdAt: now,
+            isOverloaded: false,
+            exceededMillis: 0,
+        };
 
-            // Compare with previous snapshot
+        const previousSnapshot = this.eventLoopSnapshots[this.eventLoopSnapshots.length - 1];
+        if (previousSnapshot) {
             const { createdAt } = previousSnapshot;
             const delta = now - createdAt - this.eventLoopSnapshotIntervalMillis;
-            snapshot.exceededMillis = Math.max(delta - this.maxBlockedMillis, 0);
+
             if (delta > this.maxBlockedMillis) snapshot.isOverloaded = true;
-            this.eventLoopSnapshots.push(snapshot);
-        } catch (err) {
-            log.exception(err, 'Snapshotter: Event loop snapshot failed.');
-        } finally {
-            intervalCallback();
+            snapshot.exceededMillis = Math.max(delta - this.maxBlockedMillis, 0);
         }
+
+        this.eventLoopSnapshots.push(snapshot);
+        intervalCallback();
     }
 
     /**
-     * Creates a snapshot of current CPU usage.
-     * @param {Object} cpuInfoEvent
-     * @return {Promise}
+     * Creates a snapshot of current CPU usage
+     * using the Apify platform `systemInfo` event.
+     * @param {Object} systemInfo
      * @ignore
      */
-    _snapshotCpu(cpuInfoEvent) {
-        const createdAt = (new Date(cpuInfoEvent.createdAt));
+    _snapshotCpuOnPlatform(systemInfo) {
+        const { cpuCurrentUsage, isCpuOverloaded } = systemInfo;
+        const createdAt = (new Date(systemInfo.createdAt));
         this._pruneSnapshots(this.cpuSnapshots, createdAt);
+
         this.cpuSnapshots.push({
             createdAt,
-            isOverloaded: cpuInfoEvent.isCpuOverloaded,
+            isOverloaded: isCpuOverloaded,
+            usedRatio: Math.ceil(cpuCurrentUsage / 100),
         });
+    }
+
+    /**
+     * Creates a snapshot of current CPU usage
+     * using OS provided metrics.
+     * @param {Function} intervalCallback
+     * @ignore
+     */
+    _snapshotCpuOnLocal(intervalCallback) {
+        const now = new Date();
+        this._pruneSnapshots(this.eventLoopSnapshots, now);
+
+        const ticks = this._getCurrentCpuTicks();
+        const snapshot = {
+            createdAt: now,
+            isOverloaded: false,
+            ticks,
+            usedRatio: 0,
+        };
+
+        const previousSnapshot = this.cpuSnapshots[this.cpuSnapshots.length - 1];
+        if (previousSnapshot) {
+            const { ticks: prevTicks } = previousSnapshot;
+            const idleTicksDelta = ticks.idle - prevTicks.idle;
+            const totalTicksDelta = ticks.total - prevTicks.total;
+            const usedCpuRatio = 1 - (idleTicksDelta / totalTicksDelta);
+
+            if (usedCpuRatio > this.maxUsedCpuRatio) snapshot.isOverloaded = true;
+            snapshot.usedRatio = Math.ceil(usedCpuRatio);
+        }
+
+        this.cpuSnapshots.push(snapshot);
+        intervalCallback();
+    }
+
+    _getCurrentCpuTicks() { // eslint-disable-line class-methods-use-this
+        const cpus = os.cpus();
+        return cpus.reduce((acc, cpu) => {
+            const cpuTimes = Object.values(cpu.times);
+            return {
+                idle: acc.idle + cpu.times.idle,
+                total: acc.total + cpuTimes.reduce((sum, num) => sum + num),
+            };
+        }, { idle: 0, total: 0 });
     }
 
     /**
@@ -316,32 +392,27 @@ class Snapshotter {
      * @private
      */
     _snapshotClient(intervalCallback) {
-        try {
-            const now = new Date();
-            this._pruneSnapshots(this.clientSnapshots, now);
+        const now = new Date();
+        this._pruneSnapshots(this.clientSnapshots, now);
 
-            const allErrorCounts = apifyClient.stats.rateLimitErrors;
-            const currentErrCount = allErrorCounts[2]; // 2nd retry.
+        const allErrorCounts = apifyClient.stats.rateLimitErrors;
+        const currentErrCount = allErrorCounts[CLIENT_RATE_LIMIT_ERROR_RETRY_COUNT] || 0;
 
-            // Handle empty snapshots array
-            const snapshot = {
-                createdAt: now,
-                isOverloaded: false,
-                rateLimitErrorCount: currentErrCount,
-            };
-            const previousSnapshot = this.clientSnapshots[this.clientSnapshots.length - 1];
-            if (!previousSnapshot) return this.clientSnapshots.push(snapshot);
-
-            // Compare with previous snapshot
+        // Handle empty snapshots array
+        const snapshot = {
+            createdAt: now,
+            isOverloaded: false,
+            rateLimitErrorCount: currentErrCount,
+        };
+        const previousSnapshot = this.clientSnapshots[this.clientSnapshots.length - 1];
+        if (previousSnapshot) {
             const { rateLimitErrorCount } = previousSnapshot;
             const delta = currentErrCount - rateLimitErrorCount;
             if (delta > this.maxClientErrors) snapshot.isOverloaded = true;
-            this.clientSnapshots.push(snapshot);
-        } catch (err) {
-            log.exception(err, 'Snapshotter: Client snapshot failed.');
-        } finally {
-            intervalCallback();
         }
+
+        this.clientSnapshots.push(snapshot);
+        intervalCallback();
     }
 
     /**
@@ -359,6 +430,22 @@ class Snapshotter {
             else break;
         }
         snapshots.splice(0, oldCount);
+    }
+
+    /**
+     * Calculate max memory for platform or local usage.
+     * @ignore
+     */
+    async _ensureCorrectMaxMemory() {
+        if (this.maxMemoryBytes) return;
+        const { totalBytes } = await getMemoryInfo();
+        if (isAtHome()) {
+            this.maxMemoryBytes = totalBytes;
+        } else {
+            this.maxMemoryBytes = Math.ceil(totalBytes / 4);
+            // NOTE: Log as AutoscaledPool, so that users are not confused what "Snapshotter" is
+            log.info(`AutoscaledPool: Setting max memory of this run to ${Math.round(this.maxMemoryBytes / 1024 / 1024)} MB. Use the ${ENV_VARS.MEMORY_MBYTES} environment variable to override it.`); // eslint-disable-line max-len
+        }
     }
 }