+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+0.14.10 / 2019-05-24
+===================
+- Update `Puppeteer` to 1.17.0.
+- Add `idempotencyKey` parameter to `Apify.addWebhook()`.
+
+0.14.9 / 2019-05-22
+===================
+- Better logs from `AutoscaledPool` class
+- Replace `cpuInfo` Apify event with new `systemInfo` event in `Snapshotter`.
+
 0.14.8 / 2019-05-14
 ===================
 - Bump `apify-client` to 0.5.17